# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /saves:
    get:
      operationId: listSaves
      tags: [saves]
      summary: List all saves for the current user
      security:
        - clerkAuth: []
      responses:
        "200":
          description: List of saves
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Save"
    post:
      operationId: createSave
      tags: [saves]
      summary: Add a new travel save
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaveBody"
      responses:
        "201":
          description: Save created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Save"
  /saves/{id}/tag:
    post:
      operationId: tagSave
      tags: [saves]
      summary: Generate AI tags for a save
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagSaveBody"
      responses:
        "200":
          description: Tagged save
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Save"
  /saves/{id}/geocode:
    post:
      operationId: geocodeSave
      tags: [saves]
      summary: Geocode a save using AI + Nominatim
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Geocoded save
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Save"
  /saves/{id}:
    patch:
      operationId: updateSave
      tags: [saves]
      summary: Update editable fields of a save
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSaveBody"
      responses:
        "200":
          description: Updated save
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Save"
    delete:
      operationId: deleteSave
      tags: [saves]
      summary: Delete a save
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
...
  /scrape:
    post:
      operationId: scrapeUrl
      tags: [scrape]
      summary: Scrape metadata from a URL
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScrapeUrlBody"
      responses:
        "200":
          description: Scraped metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrapeResult"
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    Save:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        note:
          type: string
          nullable: true
        url:
          type: string
          nullable: true
        scrapedTitle:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        placeName:
          type: string
          nullable: true
        countryCode:
          type: string
          nullable: true
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
        tags:
          type: array
          nullable: true
          items:
            type: string
        category:
          type: string
          nullable: true
        officialLink:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - createdAt
    UpdateSaveBody:
      type: object
      properties:
        scrapedTitle:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        placeName:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
        tags:
          type: array
          nullable: true
          items:
            type: string
        category:
          type: string
          nullable: true
        officialLink:
          type: string
          nullable: true
    CreateSaveBody:
      type: object
      properties:
        note:
          type: string
          nullable: true
        url:
          type: string
          nullable: true
        scrapedTitle:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
...
    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    ScrapeUrlBody:
      type: object
      properties:
        url:
          type: string
      required:
        - url
    TagSaveBody:
      type: object
      properties:
        extractedText:
          type: string
          nullable: true
    ScrapeResult:
      type: object
      properties:
        url:
          type: string
        title:
          type: string
          nullable: true
        image:
          type: string
          nullable: true
        siteName:
          type: string
          nullable: true
        extractedText:
          type: string
          nullable: true
      required:
        - url