# lib/api-zod/src/generated/types/createSaveBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSaveBody {
  note?: string | null;
  url?: string | null;
  scrapedTitle?: string | null;
  description?: string | null;
}