# lib/api-zod/src/generated/types/save.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Save {
  id: number;
  userId: string;
  note?: string | null;
  url?: string | null;
  scrapedTitle?: string | null;
  description?: string | null;
  placeName?: string | null;
  countryCode?: string | null;
  lat?: number | null;
  lng?: number | null;
  tags?: string[] | null;
  category?: string | null;
  officialLink?: string | null;
  createdAt: Date;
}