# lib/api-zod/src/generated/types/updateSaveBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateSaveBody {
  scrapedTitle?: string | null;
  description?: string | null;
  placeName?: string | null;
  note?: string | null;
  tags?: string[] | null;
  category?: string | null;
  officialLink?: string | null;
}